test(sdk): add OpenAI fork prompt-cache check
